requirements.txt
asgiref==3.5.2
Django==4.1.1
django-autoslug==1.9.8
django-countries==7.3.2
djangorestframework==3.14.0
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.3
typing_extensions==4.3.0
